Develop

+++ ChargeLaser.h
@@ -16,15 +16,15 @@ class ChargeLaser :
 	ID3D11PixelShader* m_PixelShader;
 	ID3D11InputLayout* m_VertexLayout;
 
-	int charge_rate = 0;//チャージ時間
+	float charge_rate = 0.f;//チャージ時間
 	ChargeEffect* eff = nullptr;//チャージ用エフェクト
 public:
 	ChargeLaser(GameObject* inobject) :Wepon(inobject)
 	{
 		dmg = 10;
-		time = 0;
+		time = 0.f;
 		angle = 0.0f;
-		charge_rate = 0;
+		charge_rate = 0.f;
 		objS = NULL;
 		eff = nullptr;
 		m_VertexShader = nullptr;
